# backend/src/api/openapi-generated.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Vibecode AI Planner API",
    "version": "1.0.0",
    "description": "API documentation for Vibecode AI Planner - project and ticket management with agent orchestration"
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token obtained from /api/auth/login"
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Agent API key for agent authentication"
      }
    },
...
  },
  "security": [
    {
      "BearerAuth": []
    }
  ],
  "paths": {
...
    "/providers": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "List all providers",
        "responses": {
          "200": {
            "description": "List of providers"
          }
        }
      },
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Add a new provider",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "providerType": {
                    "type": "string"
                  },
                  "apiKey": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created"
          }
        }
      }
    },
    "/providers/{id}": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get a provider by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider details"
          }
        }
      },
      "patch": {
        "tags": [
          "Providers"
        ],
        "summary": "Update a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated"
          }
        }
      },
      "delete": {
        "tags": [
          "Providers"
        ],
        "summary": "Delete a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider deleted"
          }
        }
      }
    },
    "/providers/{id}/test": {
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Test provider connection",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connection test result"
          }
        }
      }
    },
    "/providers/{id}/directorship": {
      "patch": {
        "tags": [
          "Providers"
        ],
        "summary": "Set provider as director",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Directorship updated"
          }
        }
      }
    },
    "/providers/{id}/agents": {
      "get": {
        "tags": [
          "Providers"
        ],
        "summary": "Get agents for a provider",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents"
          }
        }
      }
    },
    "/providers/resolve": {
      "post": {
        "tags": [
          "Providers"
        ],
        "summary": "Resolve provider for a ticket",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "labels": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  },
                  "priority": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Resolved provider config"
          }
        }
      }
    },
...
  },
...
}